feat: add default layout (#1)
close #1
