+++ src/app/(home)/globals.css
@@ -1,0 +1,17 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+  --bg-primary: black;
+  --bg-card: #121212;
+  --bg-card-hover: #181818
+}
+
+body,
+html,
+:root {
+  height: 100%;
+  background-color: var(--bg-primary);
+  color-scheme: dark;
+}
