+++ src/app/(home)/page.tsx
@@ -1,0 +1,6 @@
+
+export default function Home() {
+  return (
+    <div>home</div>
+  )
+}
