+++ src/assets/fonts/fitgree.ts
@@ -1,0 +1,3 @@
+import { Figtree } from "next/font/google"
+
+export const fitgreeFont = Figtree({ subsets: ['latin'] })
