+++ tailwind.config.ts
@@ -8,11 +8,11 @@ const config: Config = {
   ],
   theme: {
     extend: {
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-        'gradient-conic':
-          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))',
-      },
+      colors: {
+        'background': 'black',
+        'card': '#121212',
+        'card-hover': '#181818'
+      }
     },
   },
   plugins: [],
